+++ packages/cubejs-playground/src/QueryBuilderV2/components/SidePanelCubeItem.tsx
@@ -166,19 +166,11 @@ export function SidePanelCubeItem({
     );
   const segments = (filterString ? shownSegments : cube?.segments || [])
     .map((s) => s.name)
-    .filter(
-      (s) =>
-        (mode === 'all' && isOpen) ||
-        showAllMembers ||
-        query?.segments?.includes(s)
-    );
+    .filter((s) => (mode === 'all' && isOpen) || showAllMembers || query?.segments?.includes(s));
 
   if (!filterString) {
     query?.dimensions?.forEach((dimension) => {
-      if (
-        !dimensions?.includes(dimension) &&
-        dimension.startsWith(`${name}.`)
-      ) {
+      if (!dimensions?.includes(dimension) && dimension.startsWith(`${name}.`)) {
         dimensions.push(dimension);
       }
     });
@@ -216,8 +208,7 @@ export function SidePanelCubeItem({
     dateRanges.remove(name);
   }
 
-  const showMembers =
-    (isOpen || mode === 'query' || isUsed || !!filterString) && !isNonJoinable;
+  const showMembers = (isOpen || mode === 'query' || isUsed || !!filterString) && !isNonJoinable;
 
   const dimensionsSection = useMemo(() => {
     return showMembers && dimensions.length ? (
@@ -242,9 +233,7 @@ export function SidePanelCubeItem({
                   if (granularity) {
                     return (
                       query?.timeDimensions?.some(
-                        (td) =>
-                          td.dimension === item.name &&
-                          td.granularity === granularity
+                        (td) => td.dimension === item.name && td.granularity === granularity
                       ) || false
                     );
                   }
@@ -252,9 +241,8 @@ export function SidePanelCubeItem({
                   return query?.dimensions?.includes(item.name) || false;
                 }}
                 isFiltered={
-                  query?.timeDimensions?.some(
-                    (td) => td.dimension === item.name && td.dateRange
-                  ) || false
+                  query?.timeDimensions?.some((td) => td.dimension === item.name && td.dateRange) ||
+                  false
                 }
                 onDimensionToggle={(dimension) => {
                   dimensionsUpdater?.toggle(dimension);
@@ -265,9 +253,7 @@ export function SidePanelCubeItem({
                   onMemberToggle?.(name);
                 }}
                 onToggleDataRange={
-                  !dateRanges.list.includes(item.name)
-                    ? addDateRange
-                    : removeDateRange
+                  !dateRanges.list.includes(item.name) ? addDateRange : removeDateRange
                 }
               />
             );
@@ -282,9 +268,8 @@ export function SidePanelCubeItem({
               filterString={filterString}
               isSelected={query?.dimensions?.includes(name) || false}
               isFiltered={
-                query?.filters?.some(
-                  (filter) => 'member' in filter && filter.member === name
-                ) || false
+                query?.filters?.some((filter) => 'member' in filter && filter.member === name) ||
+                false
               }
               onAddFilter={addFilter}
               onRemoveFilter={removeFilter}
@@ -329,9 +314,8 @@ export function SidePanelCubeItem({
               filterString={filterString}
               isSelected={query?.measures?.includes(name) || false}
               isFiltered={
-                query?.filters?.some(
-                  (filter) => 'member' in filter && filter.member === name
-                ) || false
+                query?.filters?.some((filter) => 'member' in filter && filter.member === name) ||
+                false
               }
               onAddFilter={addFilter}
               onRemoveFilter={removeFilter}
@@ -383,8 +367,7 @@ export function SidePanelCubeItem({
 
   const hasOverflow = useHasOverflow(textRef);
 
-  const noVisibleMembers =
-    !dimensions.length && !measures.length && !segments.length;
+  const noVisibleMembers = !dimensions.length && !measures.length && !segments.length;
 
   useEffect(() => {
     setShowAllMembers(false);
@@ -417,11 +400,7 @@ export function SidePanelCubeItem({
                 type="neutral"
                 size="small"
                 icon={
-                  !showAllMembers ? (
-                    <ArrowIcon direction="bottom" />
-                  ) : (
-                    <ArrowIcon direction="top" />
-                  )
+                  !showAllMembers ? <ArrowIcon direction="bottom" /> : <ArrowIcon direction="top" />
                 }
                 placeContent="start"
                 onPress={() => setShowAllMembers(!showAllMembers)}
@@ -435,9 +414,7 @@ export function SidePanelCubeItem({
         return null;
       } else if (isOpen || mode === 'query') {
         return (
-          <Block padding=".5x 0 .5x 4.5x">
-            No members{mode === 'query' ? ' selected' : ''}
-          </Block>
+          <Block padding=".5x 0 .5x 4.5x">No members{mode === 'query' ? ' selected' : ''}</Block>
         );
       }
     } else {
@@ -451,9 +428,7 @@ export function SidePanelCubeItem({
       qaVal={name}
       icon={
         isMissing ? (
-          <QuestionCircleOutlined
-            style={{ color: 'var(--danger-text-color)' }}
-          />
+          <QuestionCircleOutlined style={{ color: 'var(--danger-text-color)' }} />
         ) : type === 'cube' ? (
           <CubeIcon color="#purple" />
         ) : (
@@ -483,15 +458,9 @@ export function SidePanelCubeItem({
       onPress={() => !isMissing && !isNonJoinable && onToggle?.(!isOpen)}
     >
       <Text ref={textRef} ellipsis>
-        {filterString ? (
-          <FilteredLabel text={name} filter={filterString} />
-        ) : (
-          name
-        )}
+        {filterString ? <FilteredLabel text={name} filter={filterString} /> : name}
       </Text>
-      {description ? (
-        <ItemInfoIcon title={title} description={description} />
-      ) : undefined}
+      {description ? <ItemInfoIcon title={title} description={description} /> : undefined}
       {isPrivate ? <NonPublicIcon type="cube" /> : undefined}
     </CubeButton>
   );